replacing previous AST with the refactored one
