+++ common/scala/syntax/definition/src/main/scala/org/enso/data/Size.scala
@@ -1,6 +1,6 @@
 package org.enso.data
 
-import org.enso.syntax.text.AST
+import org.enso.syntax.text.AST.AST
 import org.enso.syntax.text.ast.meta.Pattern
 
 /** Strongly typed size for a container. */
@@ -11,7 +11,7 @@ case class Size(value: Int) extends AnyVal with Ordered[Size] {
 
 object Size {
   val Empty                           = Size(0)
-  def apply(pat: Pattern.Match): Size = Size(pat.toStream)
+  def apply(pat: Pattern.Match): Size = ??? // FIXME Size(pat.toStream)
   def apply(ast: AST): Size           = Size(ast.span)
   def apply(text: String): Size       = Size(text.length)
   def apply(ast: Shifted[AST]): Size  = Size(ast.off) + Size(ast.el)
