+++ common/scala/syntax/definition/src/main/scala/org/enso/syntax/text/package.scala
@@ -3,5 +3,5 @@ package org.enso.syntax
 import org.enso.syntax.text.{AST => ASTDef}
 
 package object text {
-  type AST = ASTDef._AST
+  type AST = ASTDef.AST
 }
