--- common/scala/syntax/definition/src/main/scala/org/enso/syntax/text/package2.scala
@@ -1,7 +1,0 @@
-package org.enso.syntax
-
-import org.enso.syntax.text2.{AST => ASTDef}
-
-package object text2 {
-  type AST = ASTDef.AST
-}
