+++ common/scala/syntax/specialization/src/test/scala/org/enso/syntax/text/ParserTest.scala
@@ -170,12 +170,11 @@ class ParserTest extends FlatSpec with Matchers {
   //// Text ////////////////////////////////////////////////////////////////////
   //////////////////////////////////////////////////////////////////////////////
 
-  import Text.Segment.implicits.txtFromString
-
+  import Shape.SegmentPlain.txtFromString
   def line(s: String, empty: Int*) =
-    Text.Block.Line(empty.to[List], List(txtFromString[AST](s)))
+    Shape.TextBlockLine(empty.to[List], List(txtFromString[AST](s)))
   def line(segment: AST.Text.Segment.Fmt, empty: Int*) =
-    Text.Block.Line(empty.to[List], List(segment))
+    Shape.TextBlockLine(empty.to[List], List(segment))
 
   "'"    ?= Text.Unclosed()
   "''"   ?= Text()
@@ -200,14 +199,19 @@ class ParserTest extends FlatSpec with Matchers {
   "\"a\"\"" ?= Text.Unclosed.Raw("a") $ Text.InvalidQuote("\"\"")
   "\"'\""   ?= Text.Raw("'")
 
-  "\"\"\" \n\n X\n\n Y"    ?= Text.Raw(1, 0, line(" X", 0), line(" Y", 0))
-  "a \"\"\"\n\n\n X\n\n Y" ?= "a" $_ Text.Raw(0, 1, line("X", 0, 0), line("Y", 0))
+  "\"\"\" \n\n X\n\n Y" ?= Text.Raw(1, 0, line(" X", 0), line(" Y", 0))
+  "a \"\"\"\n\n\n X\n\n Y" ?= "a" $_ Text.Raw(
+    0,
+    1,
+    line("X", 0, 0),
+    line("Y", 0)
+  )
 
   //// Escapes ////
 
   val Esc = Text.Segment.Escape
   def escape(esc: Text.Segment.Escape): Text.Segment.Fmt =
-    Text.Segment._Escape(esc)
+    Shape.SegmentEscape(esc)
 
   Text.Segment.Escape.Character.codes.foreach(
     i => s"'\\$i'" ?= Text(escape(i))
@@ -226,7 +230,7 @@ class ParserTest extends FlatSpec with Matchers {
 
   //// Interpolation ////
 
-  def expr(ast: AST) = Text.Segment._Expr(Some(ast))
+  def expr(ast: AST) = Shape.SegmentExpr(Some(ast))
 
   "'a`b`c'" ?= Text("a", expr("b"), "c")
   "'a`b 'c`d`e' f`g'" ?= {
